Update ClickHouse client to 0.8.6

See merge request https://gitlab.com/gitlab-org/gitlab/-/merge_requests/216645

Merged-by: Felipe Cardozo <fcardozo@gitlab.com>
Approved-by: Pedro Pombeiro <noreply@pedro.pombei.ro>
Approved-by: Felipe Cardozo <fcardozo@gitlab.com>

+++ spec/lib/click_house/finders/ci/finished_builds_finder_spec.rb
@@ -255,7 +255,7 @@ RSpec.describe ClickHouse::Finders::Ci::FinishedBuildsFinder, :click_house, :fre
 
         it "calculates rate correctly" do
           is_expected.not_to be_empty
-          expect(result.first.fetch("rate_of_#{status}")).to be_a(Integer)
+          expect(result.first.fetch("rate_of_#{status}")).to be_a(Float)
         end
       end
